+++ test/eventQueueOutbox.test.js
@@ -18,6 +18,8 @@ const { getOpenQueueEntries } = require("../src/runner/openEvents");
 const EventQueueGenericOutboxHandler = require("../src/outbox/EventQueueGenericOutboxHandler");
 const { promisify } = require("util");
 
+const CUSTOM_HOOKS_SRV = "OutboxCustomHooks";
+
 cds.env.requires.NotificationServicePeriodic = {
   impl: "./outboxProject/srv/service/servicePeriodic.js",
   outbox: {
@@ -33,6 +35,16 @@ cds.env.requires.NotificationServicePeriodic = {
         checkForNextChunk: false,
       },
     },
+  },
+};
+
+cds.env.requires.OutboxCustomHooks = {
+  impl: "./outboxProject/srv/service/serviceCustomHooks.js",
+  outbox: {
+    kind: "persistent-outbox",
+    load: 60,
+    checkForNextChunk: true,
+    transactionMode: "isolated",
   },
 };
 
@@ -256,18 +268,7 @@ describe("event-queue outbox", () => {
       expect(loggerMock).not.sendFioriActionCalled();
       expect(payload).toMatchSnapshot();
       await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
-      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1])
-        .toMatchInlineSnapshot(`
-        {
-          "data": {
-            "body": "body",
-            "subject": "subject",
-            "to": "to",
-          },
-          "eventQueueId": "NotificationService",
-          "user": "testUser",
-        }
-      `);
+      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1]).toMatchSnapshot();
       expect(loggerMock.callsLengths().error).toEqual(0);
     });
 
@@ -288,18 +289,7 @@ describe("event-queue outbox", () => {
       expect(loggerMock).not.sendFioriActionCalled();
       expect(payload).toMatchSnapshot();
       await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
-      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1])
-        .toMatchInlineSnapshot(`
-        {
-          "data": {
-            "body": "body",
-            "subject": "subject",
-            "to": "to",
-          },
-          "eventQueueId": "NotificationService",
-          "user": "testUser",
-        }
-      `);
+      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1]).toMatchSnapshot();
       expect(loggerMock.callsLengths().error).toEqual(0);
     });
 
@@ -321,18 +311,7 @@ describe("event-queue outbox", () => {
       expect(loggerMock).not.sendFioriActionCalled();
       expect(payload).toMatchSnapshot();
       await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
-      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1])
-        .toMatchInlineSnapshot(`
-        {
-          "data": {
-            "body": "body",
-            "subject": "subject",
-            "to": "to",
-          },
-          "eventQueueId": "NotificationService",
-          "user": "badman",
-        }
-      `);
+      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1]).toMatchSnapshot();
       expect(loggerMock.callsLengths().error).toEqual(0);
     });
 
@@ -629,18 +608,7 @@ describe("event-queue outbox", () => {
       });
 
       await processEventQueue(tx.context, "CAP_OUTBOX", outboxedService.name);
-      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1])
-        .toMatchInlineSnapshot(`
-        {
-          "data": {
-            "body": "body",
-            "subject": "subject",
-            "to": "to",
-          },
-          "eventQueueId": "NotificationServiceOutboxedByConfigUserId",
-          "user": "dummyTestUser",
-        }
-      `);
+      expect(loggerMock.calls().info.find((log) => log[0].includes("sendFiori action triggered"))[1]).toMatchSnapshot();
       expect(loggerMock.callsLengths().error).toEqual(0);
     });
 
@@ -895,6 +863,170 @@ describe("event-queue outbox", () => {
         });
       });
     });
+
+    describe("custom hooks", () => {
+      describe("checkEventAndGeneratePayload", () => {
+        it("specific action call", async () => {
+          const service = (await cds.connect.to("OutboxCustomHooks")).tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+          const modifiedData = { ...data, to: "newValue" };
+          await service.send("action", data);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectOpen(tx, { expectedLength: 1 });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          expect(loggerMock).actionCalled("checkEventAndGeneratePayload.action", { data: modifiedData });
+          expect(loggerMock).actionCalled("action", { data: modifiedData });
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 1 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+
+        it("non specific action call", async () => {
+          const service = (await cds.connect.to("OutboxCustomHooks")).tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+          await service.send("main", data);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectOpen(tx, { expectedLength: 1 });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          expect(loggerMock).actionCalled("checkEventAndGeneratePayload", { data });
+          expect(loggerMock).actionCalled("main", { data });
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 1 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+
+        it("mixed both should be called", async () => {
+          const service = (await cds.connect.to("OutboxCustomHooks")).tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+          const dataSpecific = { to: "toSpecific", subject: "subject", body: "body" };
+          const modifiedData = { ...data, to: "newValue" };
+          await service.send("main", data);
+          await service.send("action", dataSpecific);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectOpen(tx, { expectedLength: 2 });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          expect(loggerMock).actionCalled("checkEventAndGeneratePayload", { data });
+          expect(loggerMock).actionCalled("checkEventAndGeneratePayload.action", { data: modifiedData });
+          expect(loggerMock).actionCalled("main", { data });
+          expect(loggerMock).actionCalled("action", { data: modifiedData });
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 2 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+      });
+
+      describe("clusterQueueEntries", () => {
+        it("non specific action call", async () => {
+          const service = (await cds.connect.to("OutboxCustomHooks")).tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+          await service.send("main", data);
+          await commitAndOpenNew();
+          const [eventEntry] = await testHelper.selectEventQueueAndReturn(tx, {
+            expectedLength: 1,
+            additionalColumns: "*",
+            parseColumns: true,
+          });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          eventEntry.lastAttemptTimestamp = expect.any(String);
+          expect(loggerMock).actionCalled("clusterQueueEntries", {
+            data: event2ClusterMap(eventEntry),
+          });
+          expect(loggerMock).actionCalled("main", { data: eventEntry.payload.data });
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 1 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+
+        it("specific action call", async () => {
+          const service = (await cds.connect.to("OutboxCustomHooks")).tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+          await service.send("action", data);
+          await commitAndOpenNew();
+          const [eventEntry] = await testHelper.selectEventQueueAndReturn(tx, {
+            expectedLength: 1,
+            additionalColumns: "*",
+            parseColumns: true,
+          });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          eventEntry.lastAttemptTimestamp = expect.any(String);
+          eventEntry.payload.data.to = "newValue";
+          expect(loggerMock).actionCalled("clusterQueueEntries.action", {
+            data: event2ClusterMap(eventEntry),
+          });
+          expect(loggerMock).actionCalled("action", { data: eventEntry.payload.data });
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 1 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+
+        it("specific and generic but no generic handler registered", async () => {
+          const srv = await cds.connect.to(CUSTOM_HOOKS_SRV);
+          const unboxedService = cds.unboxed(srv);
+          const service = srv.tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+
+          let handlerRegistration = {};
+          for (const index in unboxedService.handlers.on) {
+            const handler = unboxedService.handlers.on[index];
+            if (handler.on === "clusterQueueEntries") {
+              handlerRegistration = { index, handler };
+              delete unboxedService.handlers.on[index];
+            }
+          }
+
+          await service.send("action", data);
+          await service.send("main", data);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectOpen(tx, {
+            expectedLength: 2,
+          });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 2 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+          unboxedService.handlers.on[handlerRegistration.index] = handlerRegistration.handler;
+        });
+
+        it("should cluster two events of the same action", async () => {
+          const srv = await cds.connect.to(CUSTOM_HOOKS_SRV);
+          const service = srv.tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+
+          await service.send("action", data);
+          await service.send("action", data);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectOpen(tx, {
+            expectedLength: 2,
+          });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          expect(loggerMock).actionCalledTimes("action", 1);
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 2 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+
+        it("mixed generic and specific", async () => {
+          const srv = await cds.connect.to(CUSTOM_HOOKS_SRV);
+          const service = srv.tx(context);
+          const data = { to: "to", subject: "subject", body: "body" };
+
+          await service.send("action", data);
+          await service.send("main", data);
+          await commitAndOpenNew();
+          await testHelper.selectEventQueueAndExpectOpen(tx, {
+            expectedLength: 2,
+          });
+          await processEventQueue(tx.context, "CAP_OUTBOX", service.name);
+          await commitAndOpenNew();
+          expect(loggerMock).actionCalledTimes("clusterQueueEntries", 1);
+          expect(loggerMock).actionCalledTimes("clusterQueueEntries.action", 1);
+          expect(loggerMock).actionCalledTimes("action", 1);
+          expect(loggerMock).actionCalledTimes("main", 1);
+          await testHelper.selectEventQueueAndExpectDone(tx, { expectedLength: 2 });
+          expect(loggerMock.callsLengths().error).toEqual(0);
+        });
+      });
+    });
   });
 
   const commitAndOpenNew = async () => {
@@ -904,14 +1036,47 @@ describe("event-queue outbox", () => {
   };
 });
 
+const event2ClusterMap = (eventEntry) => ({
+  queueEntriesWithPayloadMap: {
+    [eventEntry.ID]: {
+      queueEntry: eventEntry,
+      payload: eventEntry.payload,
+    },
+  },
+});
+
 expect.extend({
-  sendFioriActionCalled: (lockerMock) => {
+  sendFioriActionCalled: () => {
     return {
       message: () => "sendFiori Action not called",
-      pass: lockerMock
+      pass: loggerMock
         .calls()
         .info.map((call) => call[0])
         .includes("sendFiori action triggered"),
+    };
+  },
+  actionCalledTimes: (loggerMock, actionName, count) => {
+    const calls = loggerMock.calls().info.filter((c) => c[0] === actionName);
+    return {
+      message: () =>
+        `expected number action of calls does not match! name: ${actionName}, expected: ${count}, actual: ${calls.length}`,
+      // eslint-disable-next-line jest/no-standalone-expect
+      pass: !expect(calls).toHaveLength(count),
+    };
+  },
+  actionCalled: (loggerMock, actionName, properties) => {
+    const call = loggerMock.calls().info.find((c) => c[0] === actionName);
+    if (!call) {
+      return {
+        message: () => `action not called! name: ${actionName}`,
+        pass: false,
+      };
+    }
+
+    return {
+      message: () => `action called with different parameters! name: ${actionName}`,
+      // eslint-disable-next-line jest/no-standalone-expect
+      pass: !expect(call[1]).toMatchObject(properties),
     };
   },
 });